Have nuPRISM correctly rotated but centred at 0 using NEUT input files
Seems to work, some strange events (neutrino interacts with no particles produced?) and lots of NC events.

+++ include/WCSimPrimaryGeneratorAction.hh
@@ -102,9 +102,9 @@ class WCSimPrimaryGeneratorAction : public G4VUserPrimaryGeneratorAction
         TTree* fRooTrackerTree;
         TTree* fSettingsTree;
         NRooTrackerVtx* fTmpRootrackerVtx;
-        double fNuPrismRadius;
-        double fNuBeamAng;
-        double fNuPlanePos[3];
+        float fNuPrismRadius;
+        float fNuBeamAng;
+        float fNuPlanePos[3];
 
     public:
